+++ src/App.css
@@ -2,12 +2,7 @@
   text-align: center;
 }
 
-.featured-ledgers{
-  display: grid;
-  justify-content: center;
-  grid-gap: 10px;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 300px) );
-}
+
