+++ src/App.js
@@ -1,17 +1,11 @@
 import "./App.css";
-import { Header, HeroSection, LedgerCard } from "./components";
-const featuredLedgers = [1, 2, 3];
+import { Header } from "./components";
+import { HomeRoute } from "./routes";
 function App() {
   return (
     <div className="App">
       <Header />
-      <HeroSection />
-      <h2>Featured Ledgers </h2>
-      <div className="featured-ledgers">
-        {featuredLedgers.map((ledger) => (
-          <LedgerCard />
-        ))}
-      </div>
+      <HomeRoute />
     </div>
   );
 }
