+++ src/routes/HomeRoute/index.css
@@ -1,0 +1,6 @@
+.featured-ledgers{
+  display: grid;
+  justify-content: center;
+  grid-gap: 10px;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 300px) );
+}
