+++ src/routes/index.js
@@ -1,0 +1,3 @@
+import HomeRoute from "./HomeRoute";
+
+export { HomeRoute };
